+++ src/domain/email/args.rs
@@ -1,20 +1,16 @@
-//! Module related to email CLI.
+//! Email CLI module.
 //!
-//! This module provides subcommands, arguments and a command matcher related to email.
+//! This module contains the command matcher, the subcommands and the
+//! arguments related to the email domain.
 
 use anyhow::Result;
-use clap::{self, App, Arg, ArgMatches, SubCommand};
-use himalaya_lib::email::TplOverride;
-use log::{debug, trace};
+use clap::{App, Arg, ArgMatches, SubCommand};
 
 use crate::{email, flag, folder, tpl, ui::table};
 
-const ARG_ATTACHMENTS: &str = "attachment";
 const ARG_CRITERIA: &str = "criterion";
-const ARG_ENCRYPT: &str = "encrypt";
 const ARG_HEADERS: &str = "header";
 const ARG_ID: &str = "id";
-const ARG_IDS: &str = "ids";
 const ARG_MIME_TYPE: &str = "mime-type";
 const ARG_PAGE: &str = "page";
 const ARG_PAGE_SIZE: &str = "page-size";
@@ -24,7 +20,7 @@ const ARG_REPLY_ALL: &str = "reply-all";
 const ARG_SANITIZE: &str = "sanitize";
 const CMD_ATTACHMENTS: &str = "attachments";
 const CMD_COPY: &str = "copy";
-const CMD_DEL: &str = "delete";
+const CMD_DELETE: &str = "delete";
 const CMD_FORWARD: &str = "forward";
 const CMD_LIST: &str = "list";
 const CMD_MOVE: &str = "move";
@@ -36,34 +32,31 @@ const CMD_SEND: &str = "send";
 const CMD_SORT: &str = "sort";
 const CMD_WRITE: &str = "write";
 
-type Criteria = String;
-type Encrypt = bool;
-type Folder<'a> = &'a str;
-type Page = usize;
-type PageSize = usize;
-type Query = String;
-type Sanitize = bool;
-type Raw = bool;
-type RawEmail<'a> = &'a str;
-type TextMime<'a> = &'a str;
-
-pub(crate) type All = bool;
-pub(crate) type Attachments<'a> = Vec<&'a str>;
-pub(crate) type Headers<'a> = Vec<&'a str>;
-pub(crate) type Id<'a> = &'a str;
-pub(crate) type Ids<'a> = &'a str;
+pub type All = bool;
+pub type Criteria = String;
+pub type Folder<'a> = &'a str;
+pub type Headers<'a> = Vec<&'a str>;
+pub type Id<'a> = &'a str;
+pub type Page = usize;
+pub type PageSize = usize;
+pub type Query = String;
+pub type Raw = bool;
+pub type RawEmail<'a> = &'a str;
+pub type Sanitize = bool;
+pub type TextMime<'a> = &'a str;
 
 /// Represents the email commands.
 #[derive(Debug, PartialEq, Eq)]
 pub enum Cmd<'a> {
     Attachments(Id<'a>),
     Copy(Id<'a>, Folder<'a>),
-    Delete(Ids<'a>),
-    Forward(Id<'a>, Attachments<'a>, Encrypt),
+    Delete(Id<'a>),
+    Flag(Option<flag::args::Cmd<'a>>),
+    Forward(Id<'a>, tpl::args::Headers<'a>, tpl::args::Body<'a>),
     List(table::args::MaxTableWidth, Option<PageSize>, Page),
     Move(Id<'a>, Folder<'a>),
     Read(Id<'a>, TextMime<'a>, Sanitize, Raw, Headers<'a>),
-    Reply(Id<'a>, All, Attachments<'a>, Encrypt),
+    Reply(Id<'a>, All, tpl::args::Headers<'a>, tpl::args::Body<'a>),
     Save(RawEmail<'a>),
     Search(Query, table::args::MaxTableWidth, Option<PageSize>, Page),
     Send(RawEmail<'a>),
@@ -74,48 +67,39 @@ pub enum Cmd<'a> {
         Option<PageSize>,
         Page,
     ),
-    Write(TplOverride<'a>, Attachments<'a>, Encrypt),
-
-    Flag(Option<flag::args::Cmd<'a>>),
     Tpl(Option<tpl::args::Cmd<'a>>),
+    Write(tpl::args::Headers<'a>, tpl::args::Body<'a>),
 }
 
 /// Email command matcher.
 pub fn matches<'a>(m: &'a ArgMatches) -> Result<Option<Cmd<'a>>> {
-    trace!("matches: {:?}", m);
-
     let cmd = if let Some(m) = m.subcommand_matches(CMD_ATTACHMENTS) {
-        debug!("attachments command matched");
         let id = parse_id_arg(m);
         Cmd::Attachments(id)
     } else if let Some(m) = m.subcommand_matches(CMD_COPY) {
-        debug!("copy command matched");
         let id = parse_id_arg(m);
         let folder = folder::args::parse_target_arg(m);
         Cmd::Copy(id, folder)
-    } else if let Some(m) = m.subcommand_matches(CMD_DEL) {
-        debug!("delete command matched");
-        let ids = parse_ids_arg(m);
-        Cmd::Delete(ids)
-    } else if let Some(m) = m.subcommand_matches(CMD_FORWARD) {
-        debug!("forward command matched");
+    } else if let Some(m) = m.subcommand_matches(CMD_DELETE) {
         let id = parse_id_arg(m);
-        let attachments = parse_attachments_arg(m);
-        let encrypt = parse_encrypt_flag(m);
-        Cmd::Forward(id, attachments, encrypt)
+        Cmd::Delete(id)
+    } else if let Some(m) = m.subcommand_matches(flag::args::CMD_FLAG) {
+        Cmd::Flag(flag::args::matches(m)?)
+    } else if let Some(m) = m.subcommand_matches(CMD_FORWARD) {
+        let id = parse_id_arg(m);
+        let headers = tpl::args::parse_headers_arg(m);
+        let body = tpl::args::parse_body_arg(m);
+        Cmd::Forward(id, headers, body)
     } else if let Some(m) = m.subcommand_matches(CMD_LIST) {
-        debug!("list command matched");
         let max_table_width = table::args::parse_max_width(m);
         let page_size = parse_page_size_arg(m);
         let page = parse_page_arg(m);
         Cmd::List(max_table_width, page_size, page)
     } else if let Some(m) = m.subcommand_matches(CMD_MOVE) {
-        debug!("move command matched");
         let id = parse_id_arg(m);
         let folder = folder::args::parse_target_arg(m);
         Cmd::Move(id, folder)
     } else if let Some(m) = m.subcommand_matches(CMD_READ) {
-        debug!("read command matched");
         let id = parse_id_arg(m);
         let mime = parse_mime_type_arg(m);
         let sanitize = parse_sanitize_flag(m);
@@ -123,25 +107,21 @@ pub fn matches<'a>(m: &'a ArgMatches) -> Result<Option<Cmd<'a>>> {
         let headers = parse_headers_arg(m);
         Cmd::Read(id, mime, sanitize, raw, headers)
     } else if let Some(m) = m.subcommand_matches(CMD_REPLY) {
-        debug!("reply command matched");
         let id = parse_id_arg(m);
         let all = parse_reply_all_flag(m);
-        let attachments = parse_attachments_arg(m);
-        let encrypt = parse_encrypt_flag(m);
-        Cmd::Reply(id, all, attachments, encrypt)
+        let headers = tpl::args::parse_headers_arg(m);
+        let body = tpl::args::parse_body_arg(m);
+        Cmd::Reply(id, all, headers, body)
     } else if let Some(m) = m.subcommand_matches(CMD_SAVE) {
-        debug!("save command matched");
         let email = parse_raw_arg(m);
         Cmd::Save(email)
     } else if let Some(m) = m.subcommand_matches(CMD_SEARCH) {
-        debug!("search command matched");
         let max_table_width = table::args::parse_max_width(m);
         let page_size = parse_page_size_arg(m);
         let page = parse_page_arg(m);
         let query = parse_query_arg(m);
         Cmd::Search(query, max_table_width, page_size, page)
     } else if let Some(m) = m.subcommand_matches(CMD_SORT) {
-        debug!("sort command matched");
         let max_table_width = table::args::parse_max_width(m);
         let page_size = parse_page_size_arg(m);
         let page = parse_page_arg(m);
@@ -149,21 +129,15 @@ pub fn matches<'a>(m: &'a ArgMatches) -> Result<Option<Cmd<'a>>> {
         let query = parse_query_arg(m);
         Cmd::Sort(criteria, query, max_table_width, page_size, page)
     } else if let Some(m) = m.subcommand_matches(CMD_SEND) {
-        debug!("send command matched");
         let email = parse_raw_arg(m);
         Cmd::Send(email)
-    } else if let Some(m) = m.subcommand_matches(CMD_WRITE) {
-        debug!("write command matched");
-        let attachments = parse_attachments_arg(m);
-        let encrypt = parse_encrypt_flag(m);
-        let tpl = tpl::args::parse_override_arg(m);
-        Cmd::Write(tpl, attachments, encrypt)
     } else if let Some(m) = m.subcommand_matches(tpl::args::CMD_TPL) {
         Cmd::Tpl(tpl::args::matches(m)?)
-    } else if let Some(m) = m.subcommand_matches(flag::args::CMD_FLAG) {
-        Cmd::Flag(flag::args::matches(m)?)
+    } else if let Some(m) = m.subcommand_matches(CMD_WRITE) {
+        let headers = tpl::args::parse_headers_arg(m);
+        let body = tpl::args::parse_body_arg(m);
+        Cmd::Write(headers, body)
     } else {
-        debug!("default list command matched");
         Cmd::List(None, None, 0)
     };
 
@@ -203,9 +177,7 @@ pub fn subcmds<'a>() -> Vec<App<'a, 'a>> {
             SubCommand::with_name(CMD_WRITE)
                 .about("Writes a new email")
                 .aliases(&["w", "new", "n"])
-                .args(&tpl::args::args())
-                .arg(attachments_arg())
-                .arg(encrypt_flag()),
+                .args(&tpl::args::args()),
             SubCommand::with_name(CMD_SEND)
                 .about("Sends a raw email")
                 .arg(raw_arg()),
@@ -223,15 +195,11 @@ pub fn subcmds<'a>() -> Vec<App<'a, 'a>> {
                 .aliases(&["rep", "r"])
                 .about("Answers to an email")
                 .arg(id_arg())
-                .arg(reply_all_flag())
-                .arg(attachments_arg())
-                .arg(encrypt_flag()),
+                .arg(reply_all_flag()),
             SubCommand::with_name(CMD_FORWARD)
                 .aliases(&["fwd", "f"])
                 .about("Forwards an email")
-                .arg(id_arg())
-                .arg(attachments_arg())
-                .arg(encrypt_flag()),
+                .arg(id_arg()),
             SubCommand::with_name(CMD_COPY)
                 .aliases(&["cp", "c"])
                 .about("Copies an email to the targeted folder")
@@ -242,10 +210,10 @@ pub fn subcmds<'a>() -> Vec<App<'a, 'a>> {
                 .about("Moves an email to the targeted folder")
                 .arg(id_arg())
                 .arg(folder::args::target_arg()),
-            SubCommand::with_name(CMD_DEL)
+            SubCommand::with_name(CMD_DELETE)
                 .aliases(&["del", "d", "remove", "rm"])
                 .about("Deletes an email")
-                .arg(ids_arg()),
+                .arg(id_arg()),
         ],
     ]
     .concat()
@@ -308,20 +276,6 @@ pub fn parse_criteria_arg<'a>(matches: &'a ArgMatches<'a>) -> String {
         .join(" ")
 }
 
-/// Represents the email ids argument.
-pub fn ids_arg<'a>() -> Arg<'a, 'a> {
-    Arg::with_name(ARG_IDS)
-        .help("Specifies the target email(s)")
-        .long_help("Specifies a range of emails. The range follows the RFC3501 format.")
-        .value_name("IDS")
-        .required(true)
-}
-
-/// Represents the email ids argument parser.
-pub fn parse_ids_arg<'a>(matches: &'a ArgMatches<'a>) -> &'a str {
-    matches.value_of(email::args::ARG_IDS).unwrap()
-}
-
 /// Represents the email reply all argument.
 pub fn reply_all_flag<'a>() -> Arg<'a, 'a> {
     Arg::with_name(ARG_REPLY_ALL)
@@ -368,24 +322,6 @@ fn parse_page_arg<'a>(matches: &'a ArgMatches<'a>) -> usize {
         .ok()
         .map(|page| 1.max(page) - 1)
         .unwrap_or_default()
-}
-
-/// Represents the email attachments argument.
-pub fn attachments_arg<'a>() -> Arg<'a, 'a> {
-    Arg::with_name(ARG_ATTACHMENTS)
-        .help("Adds attachment to the email")
-        .short("a")
-        .long("attachment")
-        .value_name("PATH")
-        .multiple(true)
-}
-
-/// Represents the email attachments argument parser.
-pub fn parse_attachments_arg<'a>(matches: &'a ArgMatches<'a>) -> Vec<&'a str> {
-    matches
-        .values_of(ARG_ATTACHMENTS)
-        .unwrap_or_default()
-        .collect()
 }
 
 /// Represents the email headers argument.
@@ -437,19 +373,6 @@ pub fn raw_arg<'a>() -> Arg<'a, 'a> {
 /// Represents the email raw argument parser.
 pub fn parse_raw_arg<'a>(matches: &'a ArgMatches<'a>) -> &'a str {
     matches.value_of(ARG_RAW).unwrap_or_default()
-}
-
-/// Represents the email encrypt flag.
-pub fn encrypt_flag<'a>() -> Arg<'a, 'a> {
-    Arg::with_name(ARG_ENCRYPT)
-        .help("Encrypts the email")
-        .short("e")
-        .long("encrypt")
-}
-
-/// Represents the email encrypt flag parser.
-pub fn parse_encrypt_flag<'a>(matches: &'a ArgMatches<'a>) -> bool {
-    matches.is_present(ARG_ENCRYPT)
 }
 
 /// Represents the email MIME type argument.